+++ src/types/event/RemotePressEvent.ts
@@ -1,0 +1,9 @@
+import type { CommonEventProperties } from './Event'
+
+export type RemotePressEvent = {
+  type: 'remotePressEvent'
+  data: {
+    id: string
+    clickPattern: 'singlePress' | 'doublePress' | 'longPress'
+  }
+} & CommonEventProperties